Refactor database to remove ORM layer
ORMs typically provide roughly 3 things:
 1. Not having to write raw SQL.
 2. Getting useful objects back in response, rather than raw rows.
 3. Nice things like lazy attribute lookup.

We care most about 1 and 2, and don't really care about 3 given our
simple data model. Thankfully, SQL Alchemy's SQL layer (it has the
SQL layer and an ORM layer on top of that) provides 1 and 2, albeit with
not quite as nice a syntax, so we'll just use that.