Test the correct behaviour
Regarding the behaviour in
https://github.com/thread/routemaster/blob/366b1279aa6f5d770b8cf6a0318a1b6b737dd6b5/routemaster/tests/test_validation.py#L143

I think it might be incorrect... if the label is in a different state
machine, we don't care about the state being deleted. We would care if
the label was in that state machine, but that's what the test above it
does.

I think currently the test is incorrect, and that on master it should be
failing. This correlates with the fact that I can't find any filtering
by state machine in
https://github.com/thread/routemaster/blob/366b1279aa6f5d770b8cf6a0318a1b6b737dd6b5/routemaster/validation.py#L51.

Looking at the git history I think this might just have been a
copy/paste issue and oversight.