Make `get_current_state` cope with deleted labels
Previously it emitted an `IndexError` which was unhelpful and, in
the one case which didn't already explicitly check for errors,
somewhat silenced by us using `suppress_exceptions`.

Changes the return type from `State` to `Optional[State]`, updating
usages to cope with this as needed. Also introduce specific tests
to pin down the behaviour of `get_current_state`, including in the
(unexpected) case where a label is in an invalid state.