Always propagate exceptions out of the Flask app
This behaviour is required for correctness. Previously Flask would generate its own 500 pages in response to uncaught exceptions; however, in doing so it would silence the exception. This meant that the middleware, critically including the session middleware, did not see exceptions. This was causing behaviour in which transactions were not actually rolling back after errors, leading to labels in deeply mysterious states.

Propagating the error up to gunicorn and letting gunicorn handle the generation of a 500 page, _after_ middleware, fixes this.