Make the timezone aware triggers resilient to delays
The cron processing includes doing the work of each scheduled job
inline at the point they detect they need to run. This works well
for the jobs where the scheduling is entirely handled by the
scheduler, because they set a next time to run and are then run
at some point after that time occurs.

However for the timezone aware triggers we're somewhat abusing
that mechanism to do our own timezone awareness filtering. As a
result, our reliance on being run at exactly (within a minute of)
the right time meant that the triggers were missing a significant
proportion of the actual times at which they should have run
their processing.

This change makes the triggers stateful -- aware of the last
real time at which they ran and thus able to account for those
delays. They now check all the minutes between when they last
ran and the current time, thus ensuring that delays still
result in the trigger running.

There is a side effect of this approach -- that delays may cause
a trigger to only run once where it might ideally run several
times. This is a worthwhile trade-off for now and is certainly
better than not running at all.