Use the time-range approach in the MetadataTimezoneAwareProcessor
This copies the same time-range approach used to optimise the
TimezoneAwareProcessor into the MetadataTimezoneAwareProcessor.
The performance increase here is less stark (I suspect because
iterating over all the timezones dominates), however this is likely
to still be an improvement. In any case it's beneficial to have
both processors working in the same way.