Use layer-loader to enable overridable values in configs
This uses https://pypi.org/project/layer-loader/ to allow
configurations to be separated into separate files and to have
placeholders defined which other files can easily override.

In order to support using placeholders, we need to add a new
entry in the schema. This new entry allows a root `placeholders'
key to be present in the configuration dict, though this value is
always ignored.