Allow the readme to be included implicitly
I'm hoping that this will get PyPI to use the README data from the
metadata rather than from the README.md, which will mean it will
pick up that it's markdown and so (hopefully) render it correctly.

The README files still get included in the sdist, so at worst I
think this removes a redundant mention of it in the MANIFEST.in.

The plugins have READMEs in reStructuredText anyway, so the mention
of a README.md in their MANIFEST.in was likely causing weirdness
anyway.