Add statsd plugin
We want to be able to monitor routemaster with Statsd based collectors
so this PR adds a copy (pretty much grep-replace my way through it) of
the Prometheus plugin but pushing to Statsd instead.

This uses https://pypi.org/project/statsd-python/ so it's not Datadog
specific + support tags (which https://pypi.org/project/python-statsd/
doesn't).

I've also added the basic logger plugin test (not sure the other
Prometheus specific tests are super relevant to copy) as well as docs +
inclusion in build and release.