Allow type checking imports to violate layering
Ideally we probably wouldn't do this, however that would require
a fairly substantial re-working of where various types are defined
and/or introducing rather a lot of protocols. We maybe should do
that at some point, however until then it seems reasonable to
allow type checking imports to specifically sidestep the layering
constraints.